Fix for build issue in 64 bit freebsd.
There are two build fixes. The first is in
OS::GetSharedLibraryAddresses(). In this method two automatic variables
have the same name: result. I changed one of them to addr to resolve the
issue.

The second one is in OS::GetCurrentThreadId(). In here, the code tried
to cast pthread_t (64 bit) to int (32 bit), resulting in error. Proper
fix should be return type change to long. But as workaround for now, I
modified it to ((int) (int64) (void *) pthread_self()) only for freebsd.